Update GitHub Actions workflows (#308)
* Remove Gradle upgrade workflow and plugin

This commit removes the automatic Gradle upgrade workflow and the associated `gradle-wrapper-upgrade` plugin.

The following changes were made:
- Deleted the `.github/workflows/gradle-upgrade.yml` file, which defined the GitHub Actions workflow for automatically upgrading Gradle.
- Removed the `libs.plugins.gradle.wrapper.upgrade` plugin alias from the root `build.gradle.kts` file.
- Removed the `wrapperUpgrade` configuration block from the root `build.gradle.kts` file.
- Deleted the `gradle-wrapper-upgrade` version and plugin declaration from the `gradle/libs.versions.toml` file.

* Refactor: Simplify GitHub Actions workflows

This commit streamlines the GitHub Actions workflows by:

- Removing redundant "Environment" steps that printed runner information.
- Consolidating environment variables to the top level of the `feature-pull-request.yml` workflow.
- Standardizing cache key naming for unit and instrumented tests execution data using environment variables (`UT_EXEC_DATA_CACHE_KEY`, `IT_EXEC_DATA_CACHE_KEY`).
- Ensuring test execution data caches are cleared at the end of the `code-quality` job.
- Refining the condition for the "Auto merge on dependencies update" job to trigger based on the branch name starting with `dependabot/` instead of labels and actor.
- Renaming "Compile code" step to "Compile sources" in `code-quality` job for clarity.
- Renaming "Auto merge when updating dependencies" job to "Auto merge on dependencies update" for conciseness.

* Fix: Grant write permissions for code quality workflow

This commit updates the `feature-pull-request.yml` GitHub Actions workflow.

The `code-quality` job now has `permissions: write-all` to allow it to comment on pull requests with code quality reports.

* Refactor: Ensure GitHub Actions cache is cleared after tests

This commit refactors the GitHub Actions workflow for feature pull requests to ensure that the caches for unit test and instrumentation test execution data are cleared *after* the tests have run and the data has been copied.

Previously, the cache was cleared before the tests, which could lead to stale data being used if the cache deletion failed. Moving the cache deletion step to after the data copy ensures that a fresh cache is always saved.