Target Android 16 (#315)
* Update compileSdk and targetSdk to 36

This commit updates the Android SDK versions used for compiling and targeting the application.

- `project-compileSdk` changed from "35" to "36"
- `project-targetSdk` changed from "35" to "36"

* Upgrade Java compatibility to version 21 and update Paparazzi

This commit updates the project's Java compatibility from version 17 to 21. This change is reflected in the Kotlin compiler options and Android compile options.

Additionally, the Paparazzi library has been updated from version 1.3.5 to 2.0.0-alpha02.

Key changes:
- Updated `jvmTarget` in `KotlinCompile` tasks to `JvmTarget.JVM_21`.
- Updated `sourceCompatibility` and `targetCompatibility` in `compileOptions` to `JavaVersion.VERSION_21`.
- Updated Paparazzi dependency from `1.3.5` to `2.0.0-alpha02`.

* Refactor: Use JDK version from version catalog for Kotlin and Java compilation

This commit updates the project configuration to use the JDK version specified in the `gradle/libs.versions.toml` file for both Kotlin and Java compilation.

Key changes:
- In `build-logic/conventions/src/main/kotlin/md/vnastasi/plugin/conventions/ProjectExt.kt`:
    - Modified `configureKotlin()` to set `jvmTarget` using `JvmTarget.fromTarget(libs.versions.project.jdk.get())`.
    - Modified `configureAndroid()` to set `sourceCompatibility` and `targetCompatibility` using `JavaVersion.toVersion(libs.versions.project.jdk.get())`.
- In `gradle/libs.versions.toml`:
    - Added a new `project-jdk` property and set its value to "21".

This change centralizes the JDK version configuration, making it easier to update across the project.