Update `dependencies-update-auto-merge` workflow trigger
This commit updates the `if` condition for the `dependencies-update-auto-merge` job in the `feature-pull-request.yml` workflow.

The condition is changed from:
`if: ${{ startsWith(github.ref_name, 'dependabot/') }}`
to:
`if: startsWith(github.head_ref, 'dependabot/') || startsWith(github.head_ref, 'wrapperbot/')`

This change ensures that the auto-merge job is triggered not only for branches starting with `dependabot/` but also for branches starting with `wrapperbot/`. This allows for automated merging of dependency updates originating from either dependabot or wrapperbot.