Update GitHub Actions checkout to v5.0.0 (#316)
* Bump actions/checkout from v4 to v5.0.0

Updates `actions/checkout` from v4 to v5.0.0 across multiple GitHub Actions workflows.

This version bump ensures that the workflows are using the latest features and bug fixes provided by the `actions/checkout` action.

Workflows updated:
- `.github/workflows/create-release.yml`
- `.github/workflows/feature-pull-request.yml`
- `.github/workflows/initiate-release.yml`
- `.github/workflows/update-caches.yml`
- `.github/workflows/update-gradle-wrapper.yml`

* Update `dependencies-update-auto-merge` workflow trigger

This commit updates the `if` condition for the `dependencies-update-auto-merge` job in the `feature-pull-request.yml` workflow.

The condition is changed from:
`if: ${{ startsWith(github.ref_name, 'dependabot/') }}`
to:
`if: startsWith(github.head_ref, 'dependabot/') || startsWith(github.head_ref, 'wrapperbot/')`

This change ensures that the auto-merge job is triggered not only for branches starting with `dependabot/` but also for branches starting with `wrapperbot/`. This allows for automated merging of dependency updates originating from either dependabot or wrapperbot.